move TFX XML serializer classes into a seperate file
